requirements.txt
requests==2.31.0
pydantic==2.8.2
apache-airflow==2.9.3
pytest==8.3.2
